+++ src/components/atomic-crm/dashboard/Dashboard.tsx
@@ -1,47 +1,21 @@
 import { useGetList } from "ra-core";
 
-import type { Contact, ContactNote } from "../types";
 import { DashboardActivityLog } from "./DashboardActivityLog";
-import { DashboardStepper } from "./DashboardStepper";
 import { DealsChart } from "./DealsChart";
-import { HotContacts } from "./HotContacts";
 import { CompletedTasksList } from "./CompletedTasksList";
 import { TasksList } from "./TasksList";
 import { Welcome } from "./Welcome";
 
 export const Dashboard = () => {
-  const {
-    data: dataContact,
-    total: totalContact,
-    isPending: isPendingContact,
-  } = useGetList<Contact>("contacts", {
-    pagination: { page: 1, perPage: 1 },
-  });
-
-  const { total: totalContactNotes, isPending: isPendingContactNotes } =
-    useGetList<ContactNote>("contact_notes", {
-      pagination: { page: 1, perPage: 1 },
-    });
-
-  const { total: totalDeal, isPending: isPendingDeal } = useGetList<Contact>(
+  const { total: totalDeal, isPending: isPendingDeal } = useGetList(
     "deals",
     {
       pagination: { page: 1, perPage: 1 },
     },
   );
 
-  const isPending = isPendingContact || isPendingContactNotes || isPendingDeal;
-
-  if (isPending) {
+  if (isPendingDeal) {
     return null;
-  }
-
-  if (!totalContact) {
-    return <DashboardStepper step={1} />;
-  }
-
-  if (!totalContactNotes) {
-    return <DashboardStepper step={2} contactId={dataContact?.[0]?.id} />;
   }
 
   return (
@@ -49,7 +23,6 @@ export const Dashboard = () => {
       <div className="md:col-span-3">
         <div className="flex flex-col gap-4">
           {import.meta.env.VITE_IS_DEMO === "true" ? <Welcome /> : null}
-          <HotContacts />
         </div>
       </div>
       <div className="md:col-span-6">