+++ src/components/atomic-crm/tasks/TasksIterator.tsx
@@ -4,11 +4,9 @@ import { useListContext } from "ra-core";
 import { Task } from "./Task";
 
 export const TasksIterator = ({
-  showContact,
   className,
   showCompleted,
 }: {
-  showContact?: boolean;
   className?: string;
   showCompleted?: boolean;
 }) => {
@@ -31,7 +29,7 @@ export const TasksIterator = ({
   return (
     <div className={`space-y-4 md:space-y-2 ${className || ""}`}>
       {tasks.map((task) => (
-        <Task task={task} showContact={showContact} key={task.id} />
+        <Task task={task} key={task.id} />
       ))}
     </div>
   );